+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2026-03-18
+
+### Changed
+
+- Attune 1 card → Attune up to Intellect cards in a single Attune action.
+
 ## 2026-03-11
 
 - **Arcane Power** / **Arcane Requirement** → **Attunement** / **Attunement Requirement**. Updated glossary definitions and rules text to match.